java_migration_tool/requirements.txt
typer>=0.9.0
rich>=13.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=2.1.1 